case study_week2_what's really warming the world?
For the second week reader contribution, I found a good data visualization example analyzing the reasons causing global warming.
In this case study, it first claimed the background story and the analytical questions clearly.
Then it analyzed each different factors separately using both verbal explanations and dynamic graphics to compare with the observed temperature movements.
After that, it combined all the dynamic graphics into one and made the results more straightforward in terms of comparisons.
In the end, the authors also provided more detailed methodology explanations with dataset sources to support the results shown above.
Overall, I think this case study is straightforward, easy to understand but also with enough information shown on each graphics, which I think can be a good example to share with the class.