issue #889 fixed
#889 Replaced a missing reference with a working one

Last week while I was reading the reader, found a link to an article on D3.js by Cabot Solutions not working. I found the one that's very similar (if not the same) and replaced it in the book.bib file so that the article is not accessible.

ps: thanks Maria for pointing out that I committed too much changes!