Updating Outliers Graphs
The graphs that talk about outliers were duplicated. I deleted the png file that is making the graph show up twice because the R code already generates the same graph. I also deleted the plot on line 27 because it doens't add to the conversation of outliers since it is the same graph with the axis swapped.