Fixed inconsistent format in 6.2
Fixed issue $1353 by adding a colon after each title. Fixed grammatical mistakes by replacing the "such as..." with "For instance" to start a new sentence after the colon. Turned the "no coding", "some coding", and "more coding" part into sub-bullets for "Choosing the tool for the project" and align the format with the sub-bullets above.